+++ apps/backend/API/Api/Models/NotificationQueryOptions.cs
@@ -6,8 +6,14 @@ namespace API.Api.Models
     {
         public int? PageNumber { get; set; }
         public int? PageSize { get; set; }
+        public bool? IsDeleted { get; set; }
+        public bool? IsAudited { get; set; }
         public byte[]? UuidBytes { get; set; }
+        public byte[]? SenderUuidBytes { get; set; }
         public NotificationType? Type { get; set; }
         public NotificationReceiverType? ReceiverType { get; set; }
+        public NotificationSenderType? ReceiverSenderType { get; set; }
+        public DateTime? StartTime { get; set; }
+        public DateTime? EndTime { get; set;}
     }
 }