+++ apps/backend/API/Application/Interfaces/ICurrentService.cs
@@ -1,0 +1,13 @@
+﻿using API.Domain.Enums;
+
+namespace API.Application.Interfaces
+{
+    public interface ICurrentService
+    {
+        bool IsAuthenticated { get; }
+        byte[]? CurrentUuid { get; }
+        string? CurrentName { get; }
+        CurrentType CurrentType { get; }
+
+    }
+}