+++ apps/backend/API/Application/Interfaces/INotificationService.cs
@@ -5,8 +5,8 @@ namespace API.Application.Interfaces
 {
     public interface INotificationService
     {
-        Task<bool> AddNotification (Notification notification);     //用于System添加无ReceiverUuid的群发Notification，以及Merchant添加对用户个体的Notification
-        Task<bool> AddBatchNotification(List<Notification> notifications);      //用于Merchant添加对于有限个数个用户的群发
+        Task<bool> AddNotification (Notification notification);     //用于添加Notification，群发或者条件发送用Delivery实现
+        Task<List<Notification>> GetNotifications (NotificationQueryOptions notificationQueryOptions);      //条件查询所需的Notifications
         Task<bool> RemoveNotificationManually(NotificationDeleteOptions notificationDeleteOptions);      //用于手动删除Notification
         Task<bool> RemoveNotificationAutomatically();       //用于自动删除out of time的Notification
     }