+++ apps/backend/API/Common/Helpers/JwtHelper.cs
@@ -15,14 +15,21 @@ namespace API.Common.Helpers
             _settings = settings;
         }
 
-        public string GenerateToken(string userId, string userRole, Guid uuid)
+        public string GenerateToken(string? openId, Guid uuid, string userRole, string? name)
         {
-            var claims = new[]
+            var claims = new List<Claim>
             {
-            new Claim(ClaimTypes.NameIdentifier, userId),
-            new Claim(ClaimTypes.Role, userRole),
-            new Claim("uuid",uuid.ToString())
-        };
+                new Claim(ClaimTypes.NameIdentifier, uuid.ToString()),
+                new Claim(ClaimTypes.Role, userRole)
+            };
+            if (!string.IsNullOrEmpty(openId))          //如果是微信小程序登录会有这个，但是因为如果Claim中有NULL值会报错，所以要拎出来写一个判断，防止报错
+            {
+                claims.Add(new Claim("OpenId", openId));
+            }
+            if(!string.IsNullOrEmpty(name))
+            {
+                claims.Add(new Claim(ClaimTypes.Name, name));           //如果有名字则add，如果没有则无视，理由同上
+            }
 
             var key = new SymmetricSecurityKey(Encoding.UTF8.GetBytes(_settings.SecretKey));
             var creds = new SigningCredentials(key, SecurityAlgorithms.HmacSha256);