+++ apps/backend/API/Domain/Entities/Models/Notification.cs
@@ -14,10 +14,6 @@ public partial class Notification
     [MaxLength(16)]
     public byte[] NotificationUuid { get; set; } = null!;
 
-    [Column("notification_receiveuuid")]
-    [MaxLength(16)]
-    public byte[]? NotificationReceiveuuid { get; set; }
-
     [Column("notification_title")]
     [StringLength(255)]
     public string NotificationTitle { get; set; } = null!;
@@ -29,8 +25,8 @@ public partial class Notification
     [Column("notification_type", TypeName = "enum('order','system','activity')")]
     public string NotificationType { get; set; } = null!;
 
-    [Column("notification_time", TypeName = "datetime")]
-    public DateTime NotificationTime { get; set; }
+    [Column("notification_starttime", TypeName = "datetime")]
+    public DateTime NotificationStarttime { get; set; }
 
     [Column("notification_isdeleted")]
     public bool NotificationIsdeleted { get; set; }
@@ -50,4 +46,10 @@ public partial class Notification
 
     [Column("notification_isaudited")]
     public bool NotificationIsaudited { get; set; }
+
+    [Column("notification_createdat", TypeName = "datetime")]
+    public DateTime NotificationCreatedat { get; set; }
+
+    [InverseProperty("DeliveryNotificationuu")]
+    public virtual ICollection<Delivery> Deliveries { get; set; } = new List<Delivery>();
 }