+++ apps/backend/API/Program.cs
@@ -69,6 +69,9 @@ namespace API
 
             builder.Services.AddLogging();
 
+            builder.Services.AddSingleton<IHttpContextAccessor, HttpContextAccessor>();
+            builder.Services.AddScoped<ICurrentService, CurrentService>();
+
             builder.Services.AddScoped<IAuthService, AuthService>();
             builder.Services.AddScoped<IUserRepository, UserRepository>();
             builder.Services.AddScoped<IUserService, UserService>();