+++ docs/CHANGES.md
@@ -1,6 +1,20 @@
 #Changelog
 这是一个开发日志，用于记录开发进度以及重要变更
 
+##[0.0.1] - 2025-6-27
+修改了部分数据结构，添加了Notification的对应Delivery表。
+完善了部分Notification的Service。
+修改了部分JWT的Claim设置。
+修改了部分AuthService。
+
+###Added
+ - 新增了NotificationService的AddNotification
+
+###Change
+ - 更变了数据库中的数据结构
+ - 更变了JWT的Claim设置
+ - 更变了AuthService与JWT对应的部分
+
 ##[0.0.0] - 2025-6-27
 将2025-6-27日前的项目初始化，已经完成了数据库结构的初步设置，后端框架的完善，自定义简要Log以及Notification的业务已经完善。